Run spark integration tests in a single suite
