+++ src/rest-test/java/edu/kit/ipd/crowdcontrol/objectservice/rest/resources/CalibrationResourceIntegrationTest.java
@@ -1,31 +1,16 @@
 package edu.kit.ipd.crowdcontrol.objectservice.rest.resources;
 
 import com.mashape.unirest.http.exceptions.UnirestException;
-import edu.kit.ipd.crowdcontrol.objectservice.Main;
-import edu.kit.ipd.crowdcontrol.objectservice.config.ConfigException;
-import edu.kit.ipd.crowdcontrol.objectservice.proto.*;
-import org.junit.AfterClass;
-import org.junit.BeforeClass;
+import edu.kit.ipd.crowdcontrol.objectservice.proto.Calibration;
+import edu.kit.ipd.crowdcontrol.objectservice.proto.CalibrationList;
 import org.junit.Test;
-import spark.Spark;
 
 import java.io.IOException;
 import java.lang.reflect.InvocationTargetException;
 
 import static org.junit.Assert.*;
-import static org.junit.Assert.assertSame;
 
 public class CalibrationResourceIntegrationTest extends ResourceIntegrationTest {
-    @BeforeClass
-    public static void setUp() throws IOException, ConfigException {
-        Main.main(null);
-    }
-
-    @AfterClass
-    public static void tearDown() {
-        Spark.stop();
-    }
-
     @Test
     public void test() throws NoSuchMethodException, UnirestException, IllegalAccessException, IOException, InvocationTargetException {
         CalibrationList list = httpGet("/calibrations", CalibrationList.class);
