+++ src/rest-test/java/edu/kit/ipd/crowdcontrol/objectservice/rest/resources/ResourceIntegrationTest.java
@@ -4,14 +4,36 @@ import com.google.protobuf.Message;
 import com.mashape.unirest.http.HttpResponse;
 import com.mashape.unirest.http.Unirest;
 import com.mashape.unirest.http.exceptions.UnirestException;
+import edu.kit.ipd.crowdcontrol.objectservice.Main;
+import edu.kit.ipd.crowdcontrol.objectservice.config.ConfigException;
+import org.junit.AfterClass;
+import org.junit.BeforeClass;
+import org.junit.runner.RunWith;
+import org.junit.runners.Suite;
+import spark.Spark;
 
 import java.io.IOException;
 import java.io.InputStream;
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
 
-public abstract class ResourceIntegrationTest {
+@RunWith (Suite.class)
+@Suite.SuiteClasses({
+        CalibrationResourceIntegrationTest.class,
+        TemplateResourceIntegrationTest.class
+})
+public class ResourceIntegrationTest {
     protected static final String ORIGIN = "http://localhost:4567";
+
+    @BeforeClass
+    public static void setUp() throws IOException, ConfigException {
+        Main.main(null);
+    }
+
+    @AfterClass
+    public static void tearDown() {
+        Spark.stop();
+    }
 
     protected <T extends Message> T httpGet(String path, Class<T> type) throws UnirestException, NoSuchMethodException, InvocationTargetException, IllegalAccessException, IOException {
         HttpResponse<InputStream> response = Unirest.get(ORIGIN + path)
